Make display styles aware of report types.
