+++ stimela/display/styles/kube.py
@@ -9,6 +9,8 @@ from rich.progress import Progress
 from rich.table import Column, Table
 from rich.text import Text
 
+from stimela.monitoring.kube import KubeReport
+
 from .base import DisplayStyle
 from .elements import status_element, timer_element
 
@@ -131,7 +133,7 @@ class KubeDisplay(DisplayStyle):
     def update(
         self,
         task_info: TaskInformation,
-        report: object,
+        report: KubeReport,
     ):
         """Updates the progress elements using the provided values.
 
@@ -151,36 +153,32 @@ class KubeDisplay(DisplayStyle):
             # why the command has square brackets in the first place.
             self.task_command.update(self.task_command_id, value=f"{(task_info.command or '--').strip('([])')}")
 
-        if report:
-            self.kube_connection_status.update(self.kube_connection_status_id, value=f"{report.connection_status}")
+        # Require a KubeReport to update all fields.
+        if not isinstance(report, KubeReport):
+            return
 
-            cpu_percent = report.total_cores * 100
+        self.kube_connection_status.update(self.kube_connection_status_id, value=f"{report.connection_status}")
 
-            self.kube_core_usage.update(self.kube_core_usage_id, value=f"[green]{cpu_percent:2.1f}[/green]%")
+        cpu_percent = report.total_cores * 100
 
-            max_cpu_percent = max(cpu_percent, self.task_maxima["kube_peak_core_usage"])
-            self.task_maxima["kube_peak_core_usage"] = max_cpu_percent
-            self.kube_peak_core_usage.update(
-                self.kube_peak_core_usage_id, value=f"[green]{max_cpu_percent:2.1f}[/green]%"
-            )
+        self.kube_core_usage.update(self.kube_core_usage_id, value=f"[green]{cpu_percent:2.1f}[/green]%")
 
-            self.kube_ram_usage.update(self.kube_ram_usage_id, value=f"[green]{report.total_memory}[/green]GB")
+        max_cpu_percent = max(cpu_percent, self.task_maxima["kube_peak_core_usage"])
+        self.task_maxima["kube_peak_core_usage"] = max_cpu_percent
+        self.kube_peak_core_usage.update(self.kube_peak_core_usage_id, value=f"[green]{max_cpu_percent:2.1f}[/green]%")
 
-            max_ram = max(report.total_memory, self.task_maxima["kube_peak_ram_usage"])
-            self.task_maxima["kube_peak_ram_usage"] = max_ram
-            self.kube_peak_ram_usage.update(self.kube_peak_ram_usage_id, value=f"[green]{max_ram}[/green]GB")
+        self.kube_ram_usage.update(self.kube_ram_usage_id, value=f"[green]{report.total_memory}[/green]GB")
 
-            self.pending_pods.update(self.pending_pods_id, value=f"[yellow]{report.pending_pods}[/yellow]")
+        max_ram = max(report.total_memory, self.task_maxima["kube_peak_ram_usage"])
+        self.task_maxima["kube_peak_ram_usage"] = max_ram
+        self.kube_peak_ram_usage.update(self.kube_peak_ram_usage_id, value=f"[green]{max_ram}[/green]GB")
 
-            self.running_pods.update(self.running_pods_id, value=f"[green]{report.running_pods}[/green]")
-
-            self.terminating_pods.update(self.terminating_pods_id, value=f"[blue]{report.terminating_pods}[/blue]")
-
-            self.successful_pods.update(self.successful_pods_id, value=f"[green]{report.successful_pods}[/green]")
-
-            self.failed_pods.update(self.failed_pods_id, value=f"[red]{report.failed_pods}[/red]")
-
-            self.stateless_pods.update(self.stateless_pods_id, value=f"[red]{report.stateless_pods}[/red]")
+        self.pending_pods.update(self.pending_pods_id, value=f"[yellow]{report.pending_pods}[/yellow]")
+        self.running_pods.update(self.running_pods_id, value=f"[green]{report.running_pods}[/green]")
+        self.terminating_pods.update(self.terminating_pods_id, value=f"[blue]{report.terminating_pods}[/blue]")
+        self.successful_pods.update(self.successful_pods_id, value=f"[green]{report.successful_pods}[/green]")
+        self.failed_pods.update(self.failed_pods_id, value=f"[red]{report.failed_pods}[/red]")
+        self.stateless_pods.update(self.stateless_pods_id, value=f"[red]{report.stateless_pods}[/red]")
 
 
 class SimpleKubeDisplay(DisplayStyle):
@@ -259,7 +257,7 @@ class SimpleKubeDisplay(DisplayStyle):
     def update(
         self,
         task_info: TaskInformation,
-        report: object,
+        report: KubeReport,
     ):
         """Updates the progress elements using the provided values.
 
@@ -277,19 +275,15 @@ class SimpleKubeDisplay(DisplayStyle):
             # why the command has square brackets in the first place.
             self.task_command.update(self.task_command_id, value=f"{(task_info.command or '--').strip('([])')}")
 
-        if report:
-            self.kube_connection_status.update(self.kube_connection_status_id, value=f"{report.connection_status}")
+        # Require a KubeReport to update all fields.
+        if not isinstance(report, KubeReport):
+            return
 
-            cpu_percent = report.total_cores * 100
-
-            self.kube_core_usage.update(self.kube_core_usage_id, value=f"[green]{cpu_percent:2.1f}[/green]%")
-
-            self.kube_ram_usage.update(self.kube_ram_usage_id, value=f"[green]{report.total_memory}[/green]GB")
-
-            self.pending_pods.update(self.pending_pods_id, value=f"[yellow]{report.pending_pods}[/yellow]")
-
-            self.running_pods.update(self.running_pods_id, value=f"[cyan]{report.running_pods}[/cyan]")
-
-            self.successful_pods.update(self.successful_pods_id, value=f"[green]{report.successful_pods}[/green]")
-
-            self.failed_pods.update(self.failed_pods_id, value=f"[red]{report.failed_pods}[/red]")
+        self.kube_connection_status.update(self.kube_connection_status_id, value=f"{report.connection_status}")
+        cpu_percent = report.total_cores * 100
+        self.kube_core_usage.update(self.kube_core_usage_id, value=f"[green]{cpu_percent:2.1f}[/green]%")
+        self.kube_ram_usage.update(self.kube_ram_usage_id, value=f"[green]{report.total_memory}[/green]GB")
+        self.pending_pods.update(self.pending_pods_id, value=f"[yellow]{report.pending_pods}[/yellow]")
+        self.running_pods.update(self.running_pods_id, value=f"[cyan]{report.running_pods}[/cyan]")
+        self.successful_pods.update(self.successful_pods_id, value=f"[green]{report.successful_pods}[/green]")
+        self.failed_pods.update(self.failed_pods_id, value=f"[red]{report.failed_pods}[/red]")
